unsigned int to sizze_t

+++ ft_bzero.c
@@ -6,11 +6,13 @@
 /*   By: crmanzan <crmanzan@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/15 16:10:45 by crmanzan          #+#    #+#             */
-/*   Updated: 2023/09/15 17:31:36 by crmanzan         ###   ########.fr       */
+/*   Updated: 2023/09/16 15:25:07 by crmanzan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	ft_bzero(void *s, unsigned int n)
+#include <stdlib.h>
+
+void	ft_bzero(void *s, size_t n)
 {
 	unsigned char	*c;
