+++ ft_memcpy.c
@@ -6,14 +6,15 @@
 /*   By: crmanzan <crmanzan@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/14 19:45:53 by crmanzan          #+#    #+#             */
-/*   Updated: 2023/09/15 16:10:12 by crmanzan         ###   ########.fr       */
+/*   Updated: 2023/09/16 15:22:13 by crmanzan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <string.h>
 #include <stdio.h>
+#include <stdlib.h>
 
-void	*ft_memcpy(void *dst, const void *src, unsigned int n)
+void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
 	unsigned const char	*s;
 	unsigned char		*d;
